Phase_1_Submission/requirements.txt
customtkinter==5.2.2
Pillow>=10.4.0
ExifRead==3.0.0
PyMuPDF==1.25.1
